elevate proto message creation/marshalling
previously, we were creating the protobuf messages for imports, and then
marshalling them to bytes on a node-by-node basis. The data, however, was the
same for each node, so in the case of replication > 1 this was duplicating a lot
of work. By doing it ahead of time, we can improve performance, and (the
original motivation), we can log the import just once instead of once per node.