add ability to add to lru cache without eviction, fix ingest bug
go-pilosa keeps an LRU cache of the row key to id mapping so that it
doesn’t have to look it all up from Pilosa every time. It first checks
the cache to see which keys it needs to look up, then looks up from
Pilosa and adds to the cache. Unfortunately, it was possible for the
newly added keys to evict ones which were previously in the
cache. When it went to use them, it couldn’t find them and errored
out, which eventually caused ingest to deadlock (when using PDK).