discard slices rather than reslicing them
In a previous version of this, I thoughtfully resliced slices
to [:0] rather than allocating. That's a great idea, except that
you don't necessarily ever see a given shard again. If you
don't see that shard again, or if you don't get as many items
in it, all the old record pointers are still sitting there in
reachable parts of the slice's backing store, preventing them
from ever being garbage collected.