support for creation of decimal fields, expose import code
go-pilosa can now create decimal type fields — this functionality is
currently only available in Molecula's Pilosa fork.

expose import code so that the batch importer can use it to import to
mutex fields. Allow mutex fields in the batch importer.

For some reason go-pilosa was still generating deprecated "Range"
queries instead of "Row", so put a stop to that.