fix deadlock, catch empty translate keys, support []uint64 in batch
There was a deadlock which @tgruben found in detectClusterChanges
where it could acquire a lock and never release it if there were no
indexes or shards.

TranslateKeys requests fail if the slice of keys is empty. We now test
for this condition and return an empty slice of uint64 without hitting
Pilosa.

In importbatch, if a nil is encountered and the field is not an
integer or decimal field, there is no way of knowing if this value
will be using []string normally, or string or uint64. Instead of
unconditionally appending to rowIDs, we wait until we know which value
type this will be (as soon as we receive a non-nil value). Then we
nil-extend rowIDs for this field if necessary.

We now support []uint64 in addition to []string in importbatch — this
is analagous to supporting strings and uint64, and they can be
intermixed in the same value field.

Also changed the way that indexes and fields are encoded to strings
because a test was breaking due to the addition of a mutex to the
Index. I don't think this affects much of anything.