fix single batch clear bug, add logging, change record ID translation
a lot of this logging should probably be tracing

the toTranslateID change became necessary because although IDs were
coming in roughly sorted, they were getting randomized by the
map. This was fine until the translated IDs crossed a slice boundary,
and then 1000s of requests were being generated for a single batch as
the value importer flipped back and forth between slices.

the bug fix is explained in a comment, but basically makeFragments
iterated over b.rowIDs and assumed that anything in clearRowIDs would
also be in rowIDs. In this particular case, nothing had yet been added
to rowIDs, so the clear was skipped.